Refactor as a service and migrate `FetchOrigin`
